repository/default-agent/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-openai
beautifulsoup4 
WeasyPrint
requests
openai 
pinecone-client 
langchain-text-splitters 
tiktoken
unidecode